add version page, move off Status to Settings
